requirements.txt
Flask
Authlib
python-dotenv
requests
twilio
SpeechRecognition
sounddevice
scipy
numpy
pydub